+++ src/app/Repositories/Transfer/TransferRepository.php
@@ -91,13 +91,11 @@ class TransferRepository implements TransferRepositoryInterface
      * it deducts the amount from the user's balance and persists the change.
      *
      * @param User $user  The user whose balance will be debited.
-     * @param float $amount  The amount to be debited.
-     *
-     * @throws InsufficientFundsException  If the user's balance is insufficient.
+     * @param int $amount  The amount to be debited.
      *
      * @return User  The updated User instance with the new balance.
      */
-    public function debitUser(User $user, float $amount): User
+    public function debitUser(User $user, int $amount): User
     {
         $user->balance -= $amount;
         $this->updateUserBalance($user);
@@ -112,11 +110,11 @@ class TransferRepository implements TransferRepositoryInterface
      * and persists the change to the database.
      *
      * @param User $user  The user whose balance will be credited.
-     * @param float $amount  The amount to be credited.
+     * @param int $amount  The amount to be credited.
      *
      * @return User  The updated User instance with the new balance.
      */
-    public function creditUser(User $user, float $amount): User
+    public function creditUser(User $user, int $amount): User
     {
         $user->balance += $amount;
         $this->updateUserBalance($user);
